+++ Snippets/Snippets_3/Handlers/MyHandler.cs
@@ -8,8 +8,8 @@
     {
         public void Handle(MyMessage message)
         {
-            //do something relevant with the message
-            SomeLibrary.SomeMethod(message);
+            //do something with the message data
+            SomeLibrary.SomeMethod(message.Data);
         }
     }
