+++ Snippets/Snippets_3/Snippets_3.csproj
@@ -203,7 +203,7 @@
     <Compile Include="Container\Custom\MyCustomObjectBuilder.cs" />
     <Compile Include="Container\Custom\Usage.cs" />
     <Compile Include="Container\MyService.cs" />
-    <Compile Include="Handlers\GenericMessageHandler.cs" />
+    <Compile Include="Handlers\GenericHandler.cs" />
     <Compile Include="Handlers\MyMessage.cs" />
     <Compile Include="Handlers\MyHandler.cs" />
     <Compile Include="ImmediateDispatch\Usage.cs" />
