+++ Snippets/Snippets_4/Snippets_4.csproj
@@ -281,7 +281,7 @@
     <Compile Include="Container\Containers_Unity.cs" />
     <Compile Include="Container\Custom\MyCustomObjectBuilder.cs" />
     <Compile Include="Container\MyService.cs" />
-    <Compile Include="Handlers\GenericMessageHandler.cs" />
+    <Compile Include="Handlers\GenericHandler.cs" />
     <Compile Include="Handlers\MyHandler.cs" />
     <Compile Include="Handlers\MyMessage.cs" />
     <Compile Include="ImmediateDispatch\Usage.cs" />
