+++ Snippets/Snippets_5/Handlers/MyMessage.cs
@@ -2,5 +2,6 @@
 {
     public class MyMessage
     {
+        public string Data { get; set; }
     }
 }
