+++ Snippets/Snippets_5/Snippets_5.csproj
@@ -391,7 +391,7 @@
     <Compile Include="Errors\SecondLevel\DisableWithCode.cs" />
     <Compile Include="Features\FeatureConfiguration.cs" />
     <Compile Include="Features\MyFeature.cs" />
-    <Compile Include="Handlers\GenericMessageHandler.cs" />
+    <Compile Include="Handlers\GenericHandler.cs" />
     <Compile Include="Handlers\MyHandler.cs" />
     <Compile Include="Handlers\MyMessage.cs" />
     <Compile Include="ImmediateDispatch\Usage.cs" />
